backend/pyproject.toml
[project]
name = "voxia-backend"
version = "1.0.0"
description = "VoxIA Backend — Transcripción y análisis de reuniones con IA"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==3.2.2",
    "python-multipart==0.0.9",
    "python-dotenv==1.0.1",
    "numpy==1.26.4",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "aiofiles==24.1.0",
    "reportlab==4.2.5",
    "mlflow>=2.18.0",
    "grpcio>=1.60.0",
    "protobuf>=6.31.1",
    "setuptools",
]

[project.optional-dependencies]
ml = [
    "torch>=2.1.0",
    "openai-whisper==20240930",
    "transformers==4.44.2",
    "accelerate==0.34.2",
    "sentencepiece==0.2.0",
    "ffmpeg-python==0.2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.uv.extra-build-dependencies]
openai-whisper = ["setuptools<71"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

